services/jwt-issuer/app/requirements.txt
fastapi
uvicorn[standard]
redis
requests
python-multipart 
PyJWT==2.8.0
httpx
boto3
